+++ python/featomic/tests/clebsch_gordan/rotations.py
@@ -96,9 +96,12 @@ class WignerDReal:
         :param angles: Sequence[float], the alpha, beta, gamma Euler angles, in radians.
         """
         self.max_angular = max_angular
-        # Randomly generate Euler angles between 0 and 2 pi if none are provided
+        # Randomly generate Euler angles between if none are provided
         if angles is None:
-            angles = np.random.uniform(size=(3)) * 2 * np.pi
+            psi = 2 * np.pi * np.random.uniform()  # psi is between 0 and 2pi
+            theta = np.pi * np.random.uniform()  # theta is between 0 and pi
+            phi = 2 * np.pi * np.random.uniform()  # phi is between 0 and 2pi
+            angles = np.array([psi, theta, phi])
         self.angles = angles
         self.rotation = cartesian_rotation(angles)